added chisq

+++ fit_disk.py
@@ -2,7 +2,9 @@ import os
 import sys
 import pdb
 import time
+import copy
 import emcee
+import mpfit
 import numpy as np
 import astropy.io.fits as fits
 import astropy.units as u
@@ -10,12 +12,13 @@ import astropy.wcs as WCS
 from scipy.ndimage import rotate as rotate_image
 from matplotlib import pyplot as plt
 
+
 def rebin(arr, shape):
     sh = shape[0], arr.shape[0]//shape[0], shape[1], arr.shape[1]//shape[1]
     return arr.reshape(sh).mean(-1).mean(1)
 
-dir = "/Users/rcooke/Work/Research/Accel/data/TW_Hya/2016.1.00440.S/science_goal.uid___A001_X889_X18e/group.uid___A001_X889_X18f/member.uid___A001_X889_X190/product/"
-#dir = "/Users/rcooke/Desktop/datatrans/"
+#dir = "/Users/rcooke/Work/Research/Accel/data/TW_Hya/2016.1.00440.S/science_goal.uid___A001_X889_X18e/group.uid___A001_X889_X18f/member.uid___A001_X889_X190/product/"
+dir = "/Users/rcooke/Desktop/datatrans/"
 #fname = "TW_Hya_contsub_CSv0-tclean.image.pbcor.copy.vmap.fits"
 fname = "test.vmap.fits"
 
@@ -32,7 +35,7 @@ world = w.wcs_pix2world(coord, 1)
 dradec = (np.pi/180.0)*abs(world[1, 1]-world[0, 1])  # Note, this is in radians
 #dra = u.arcsec * 3600.0*(world[0, 0]-world[1, 0]) * np.cos(world[0, 1]*np.pi/180.0)
 
-subgrid = 11
+subgrid = 3
 y = data[:, :, 0].flatten()
 ye = data[:, :, 1].flatten()
 x = np.zeros(y.shape)
@@ -45,12 +48,11 @@ xx, yy = np.meshgrid(xone, yone)
 
 # Set the priors
 mn_vlsr, mx_vlsr = -3.5, -2.2
-mn_Mstar, mx_Mstar = 0.6, 1.0
+mn_Mstar, mx_Mstar = 0.6/60.1, 1.0/60.1
 mn_xcen, mx_xcen = 0.0, data.shape[0]
 mn_ycen, mx_ycen = 0.0, data.shape[1]
 mn_ang_i, mx_ang_i = 0.0, np.pi/2.0
 mn_theta, mx_theta = 0.0, np.pi
-mn_dist, mx_dist = 60.0, 60.2
 
 # Some good reference papers:
 # https://arxiv.org/pdf/1801.03948.pdf
@@ -91,16 +93,17 @@ mn_dist, mx_dist = 60.0, 60.2
 # radius ==> dist * sqrt((xd-xcen)**2 + ((yd-ycen)/cos(ang_i))**2)
 # theta ==> arctan((xd-xcen)*cos(ang_i)/(yd-ycen))
 
+
 def get_model(par):
-    vlsr, Mstar, xcen, ycen, ang_i, theta, dist = par
-    dist = dist * u.pc
-    Mstar = Mstar * u.Msun
+    vlsr, Mstar, xcen, ycen, ang_i, theta = par
+    dist = 60.1 * u.pc
+    Mstar = Mstar * u.Msun / dist
     model = vlsr*np.ones((data.shape[0], data.shape[1]))
     xd = rotate_image(xx, theta*180.0/np.pi, reshape=False)
     yd = rotate_image(yy, theta*180.0/np.pi, reshape=False)
     thetapar = np.arctan2(yd-ycen, xd-xcen)
     thetapar[np.where((xd-xcen == 0) & (xd == 0))] = 0.0
-    radius = dist * dradec * np.sqrt((xd-xcen)**2 + ((yd-ycen)/np.cos(ang_i))**2)
+    radius = dradec * np.sqrt((xd-xcen)**2 + ((yd-ycen)/np.cos(ang_i))**2)
     vshift = np.sqrt(Gcons * Mstar / radius) * np.sin(ang_i) * np.cos(thetapar)
     vshift[np.where(radius == 0.0)] = 0.0
     vshrb = rebin(vshift, (data.shape[0], data.shape[1]))
@@ -117,14 +120,13 @@ def lnprior(par):
     # ang_i = angle of incidence
     # theta = angle on the observer's sky where the disk plane intersects the plane of the sky
     # dist  = Distance from observer to the disk
-    vlsr, Mstar, xcen, ycen, ang_i, theta, dist = par
+    vlsr, Mstar, xcen, ycen, ang_i, theta = par
     if mn_vlsr <= vlsr <= mx_vlsr and \
        mn_Mstar <= Mstar <= mx_Mstar and \
        mn_xcen <= xcen <= mx_xcen and \
        mn_ycen <= ycen <= mx_ycen and \
        mn_ang_i <= ang_i <= mx_ang_i and \
-       mn_theta <= theta <= mx_theta and \
-       mn_dist <= dist <= mx_dist:
+       mn_theta <= theta <= mx_theta:
         return 0.0
     return -np.inf
 
@@ -141,6 +143,7 @@ def lnprob(theta, x, y, yerr):
         return -np.inf
     return lp + lnlike(theta, x, y, yerr)
 
+
 def run_mcmc():
     # Set up the sampler.
     ndim, nwalkers = 6, 100
@@ -150,8 +153,7 @@ def run_mcmc():
                      np.random.uniform(mn_xcen, mx_xcen),
                      np.random.uniform(mn_ycen, mx_ycen),
                      np.random.uniform(mn_ang_i, mx_ang_i),
-                     np.random.uniform(mn_theta, mx_theta),
-                     np.random.uniform(mn_dist, mx_dist)]) for i in range(nwalkers)]
+                     np.random.uniform(mn_theta, mx_theta)]) for i in range(nwalkers)]
     sampler = emcee.EnsembleSampler(nwalkers, ndim, lnprob, args=(x, y, ye), threads=ndim)
 
     # Clear and run the production chain.
@@ -168,13 +170,81 @@ def run_mcmc():
     np.save("diskfit.npy", sampler.chain)
 
 
+def myfunct(p, fjac=None, x=None, y=None, err=None):
+    model = get_model(p).flatten()
+    # Non-negative status value means MPFIT should
+    # continue, negative means stop the calculation.
+    status = 0
+    return [status, (y-model)/err]
+
+
+def run_chisq():
+    #######################################
+    #          PREPARE THE FIT
+    #######################################
+
+    # Set some reasonable starting conditions
+    p0 = [-2.8, 0.8/60.1, 320.0, 335.0, 25.0*np.pi/180.0, 16.0*np.pi/180.0]
+    #p0 = [-2.75, 0.82/60.1, 330.0, 340.0, 5.0 * np.pi / 180.0, 36.0 * np.pi / 180.0]
+
+    # Set some constraints you would like to impose
+    param_base = {'value': 0., 'fixed': 0, 'limited': [1, 1], 'limits': [0., 0.]}
+
+    # Make a copy of this 'base' for all of our parameters, and set starting parameters
+    param_info = []
+    for i in range(len(p0)):
+        param_info.append(copy.deepcopy(param_base))
+        param_info[i]['value'] = p0[i]
+
+    # Put some priors on the parameters
+    param_info[0]['limits'] = [mn_vlsr, mx_vlsr]
+    param_info[1]['limits'] = [mn_Mstar, mx_Mstar]
+    param_info[2]['limits'] = [mn_xcen, mx_xcen]
+    param_info[3]['limits'] = [mn_ycen, mx_ycen]
+    param_info[4]['limits'] = [mn_ang_i, mx_ang_i]
+    param_info[5]['limits'] = [mn_theta, mx_theta]
+
+    # Now tell the fitting program what we called our variables
+    fa = {'x': xx.flatten(), 'y': y.flatten(), 'err': ye.flatten()}
+
+    #######################################
+    #  PERFORM THE FIT AND PRINT RESULTS
+    #######################################
+
+    m = mpfit.mpfit(myfunct, p0, parinfo=param_info, functkw=fa, quiet=False)
+    if (m.status <= 0):
+        print('error message = ', m.errmsg)
+    print("param: ", m.params)
+    print("error: ", m.perror)
+
+    #######################################
+    #  PERFORM THE FIT AND PRINT RESULTS
+    #######################################
+
+    # Generate the best-fitting model
+    model = get_model(m.params)
+
+    #######################################
+    #            PLOT THE DATA
+    #######################################
+
+    plt.subplot(211)
+    plt.imshow(data[:, :, 0])
+    plt.subplot(212)
+    plt.imshow(data[:, :, 0]-model)
+    plt.show()
+
+
 if __name__ == "__main__":
-    if True:
+
+    if False:
         # Do the production run!
         run_mcmc()
-    elif False:
+    elif True:
+        run_chisq()
+    elif True:
         # Run some tests
-        p0 = [-2.75, 0.82, 330.0, 340.0, 5.0*np.pi/180.0, 36.0*np.pi/180.0, 60.1]
+        p0 = [-2.75, 0.82/60.1, 330.0, 340.0, 5.0*np.pi/180.0, 36.0*np.pi/180.0]
         model = get_model(p0)
         vmap = np.zeros((model.shape[0], model.shape[1], 2))
         errs = np.random.uniform(0.01, 0.02, model.shape)
@@ -183,10 +253,11 @@ if __name__ == "__main__":
         # Save the fake data
         fname = "test.fits"
         outname = dir+fname.split(".fits")[0] + '.vmap.fits'
-        os.remove(outname)
+        if os.path.exists(outname):
+            os.remove(outname)
         header = w.to_header()
         vmaphdu = fits.PrimaryHDU(vmap, header=header)
         vmaphdu.writeto(outname)
-        plt.imshow(model)
+        plt.imshow(vmap[:, :, 0])
         plt.show()
         print("File saved:\n"+outname)
